# Conflicts:
#	camera-view/build.gradle
#	camera-view/src/androidTest/java/com/google/android/cameraview/AspectRatioInstrumentationTest.java
#	camera-view/src/androidTest/java/com/google/android/cameraview/CameraViewActions.java
#	camera-view/src/androidTest/java/com/google/android/cameraview/CameraViewMatchers.java
#	camera-view/src/androidTest/java/com/google/android/cameraview/CameraViewTest.java
#	camera-view/src/main/api14/com/google/android/cameraview/Camera1.java
#	camera-view/src/main/api21/com/google/android/cameraview/Camera2.java
#	camera-view/src/main/api9/com/google/android/cameraview/SurfaceViewPreview.java
#	camera-view/src/main/base/com/google/android/cameraview/AspectRatio.java
#	camera-view/src/main/base/com/google/android/cameraview/Size.java
#	camera-view/src/main/base/com/google/android/cameraview/SizeMap.java
#	camera-view/src/main/java/com/google/android/cameraview/CameraView.java
#	config.gradle
#	custom-view/src/main/java/ru/surfstudio/android/custom/view/placeholder/StandardPlaceHolderView.kt
#	filestorage-sample/src/main/java/ru/surfstudio/android/filestorage/sample/interactor/common/network/cache/CacheModule.kt
#	gradle.properties
#	network-sample/src/main/java/ru/surfstudio/android/network/sample/interactor/common/network/cache/CacheModule.kt
#	picture-provider/src/main/java/ru/surfstudio/android/picturechooser/ChooserPictureProvider.kt
#	picture-provider/src/main/java/ru/surfstudio/android/picturechooser/GalleryPictureProvider.kt
#	picture-provider/src/main/java/ru/surfstudio/android/picturechooser/Utils.kt
#	push/src/main/java/ru/surfstudio/android/notification/ui/notification/strategies/PushHandleStrategy.kt
#	recycler-extension/src/main/java/ru/surfstudio/android/recycler/extension/divider/DividerItemDecoration.java
#	recycler-extension/src/main/java/ru/surfstudio/android/recycler/extension/snaphelper/GravitySnapHelper.java
#	recycler-extension/src/main/java/ru/surfstudio/android/recycler/extension/sticky/layoutmanager/StickyItemPositioner.kt
#	recycler-extension/src/main/java/ru/surfstudio/android/recycler/extension/sticky/layoutmanager/StickyLayoutManager.kt
#	shared-pref-sample/src/main/java/ru/surfstudio/android/shared/pref/sample/interactor/common/network/cache/CacheModule.kt
#	template/app-injector/build.gradle
#	template/app-injector/src/main/java/ru/surfstudio/standard/app_injector/network/cache/CacheModule.kt
#	template/base/build.gradle
#	template/config.gradle
#	template/f-main/src/main/java/ru/surfstudio/standard/f_main/MainActivityView.kt

+++ network-sample/src/main/java/ru/surfstudio/android/network/sample/interactor/common/network/OkHttpModule.kt
@@ -13,18 +13,20 @@ import java.util.concurrent.TimeUnit
 class OkHttpModule {
 
     companion object {
-        private const val NETWORK_TIMEOUT = 10 //sec
+        private const val NETWORK_TIMEOUT = 10L //sec
     }
 
     @Provides
     @PerApplication
-    fun provideOkHttpClient(cacheInterceptor: SimpleCacheInterceptor,
-                                     etagInterceptor: EtagInterceptor,
-                                     httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
+    fun provideOkHttpClient(
+            cacheInterceptor: SimpleCacheInterceptor,
+            etagInterceptor: EtagInterceptor,
+            httpLoggingInterceptor: HttpLoggingInterceptor
+    ): OkHttpClient {
         return OkHttpClient.Builder().apply {
-            connectTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
-            readTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
-            writeTimeout(NETWORK_TIMEOUT.toLong(), TimeUnit.SECONDS)
+            connectTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
+            readTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
+            writeTimeout(NETWORK_TIMEOUT, TimeUnit.SECONDS)
 
             addInterceptor(cacheInterceptor)
             addInterceptor(etagInterceptor)
